requirements.txt
openai>=1.43.0
python-dotenv>=1.0.1
GitPython>=3.1.43
tqdm>=4.66.4
PyYAML>=6.0.2
